+++ src/Predicate.php
@@ -33,6 +33,16 @@ final class Predicate implements Stringable
         $this->args     = $args;
     }
 
+    /** @param mixed[] $data */
+    public static function __set_state(array $data) : self
+    {
+        return new static(
+            $data['name'],
+            $data['fragment'],
+            $data['args']
+        );
+    }
+
     public function __toString() : string
     {
         return $this->q();