+++ src/Exception/InvalidArgument.php
@@ -6,7 +6,6 @@ namespace Prismic\Exception;
 use InvalidArgumentException;
 use Prismic\Json;
 use Prismic\Value\FormField;
-use Psr\Http\Message\UriInterface;
 use function get_class;
 use function gettype;
 use function is_object;
@@ -75,15 +74,6 @@ class InvalidArgument extends InvalidArgumentException implements PrismicError
             'The form field "%s" expects an integer value but received %s',
             $field->name(),
             is_object($invalidValue) ? get_class($invalidValue) : gettype($invalidValue)
-        ));
-    }
-
-    public static function mismatchedPreviewHost(UriInterface $apiUri, UriInterface $previewUri) : self
-    {
-        return new static(sprintf(
-            'The preview url has been rejected because its host name "%s" does not match the api host "%s"',
-            $previewUri->getHost(),
-            $apiUri->getHost()
         ));
     }
 }